### 2022-09-16 #4769 @editorialbot
[PRE REVIEW]: Guitarsounds : A Python package to visualize harmonic sounds for musical instrument testing
**Submitting author:** @olivecha (<a href="http://orcid.org/0000-0003-3294-2812">Olivier Chabot</a>)
**Repository:** https://github.com/olivecha/guitarsounds
**Branch with paper.md** (empty if default branch): JOSS-Paper
**Version:** v1.1.0
