### 2022-09-16 #4769 @danielskatz
👋 @olivecha - thanks for your submission.  A few comments/questions:

1. It doesn't have any references, which is quite unusual for a JOSS paper. Is there no related work or anything that this builds on that should be cited?
2. Is this software intended for research, or for education?
