### 2022-09-16 #4769 @olivecha
Hi Daniel,

  1.  There is a citation for the librosa package which intersects with guitarsounds (see paper.bib). The package builds upon the numpy/scipy/matplotlib stack which could be cited.
  2.  It's main purpose is research but it is also used in education to introduce students to signal processing and sound analysis for musical instrument desing.

Hope this answer your questions,

Thanks,

Olivier

Get Outlook for iOS<https://aka.ms/o0ukef>
