### 2022-09-16 #4769 @danielskatz
Re 1, there is no reference that builds in the paper, so perhaps you aren't putting the citation into the text that would trigger this.  Please see the [sample paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).

Re 2, can you perhaps add a little more to the paper that explains the research use?

Please feel free to make changes to your .bib or .md files, then use the commands `@editorialbot check references` to check references and `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
