+++ paper/paper.bib
@@ -1,9 +1,9 @@
-@inproceedings{McFee:2015,
+@inproceedings{Librosa,
+	       Title = {librosa: {Audio} and {Music} {Signal} {Analysis} in {Python}},
+	       doi = {10.25080/Majora-7b98e3ed-003},
 	       url = {https://conference.scipy.org/proceedings/scipy2015/brian_mcfee.html},
 	       Author = {{McFee}, B. and {Raffel}, C. and {Liang}, D. and {Ellis}, D. and {McVicar}, M. and {Battenberg}, E. and {Nieto}, O.},
-	       Title = {librosa: {Audio} and {Music} {Signal} {Analysis} in {Python}},
 	       Year = {2015},
-	       doi = {10.25080/Majora-7b98e3ed-003},
 	       pages = {18--24},
 	       }
 
