+++ paper/paper.md
@@ -60,7 +60,7 @@ plt.gcf().savefig('peak_comparison')
 
 # Statement of need
 
-Guitarsounds was written to meet the needs of the Bruant lutherie school, more precisely as a tool to visualize and compare the sounds of different guitar designs based on arbitrary sound features. The guitarsounds API serves as a introduction to programming for data analysis for the school's students and the GUI allows visualizing sound features even for users with minimal affinity with computers, as such is the case for some students. More over, the guitarsounds package is used in the teaching activities as a tool to visualize the physical phenomenons involved in the sound produced by guitars, such as the hemlotz cavity frequency of the instrument. The GUI was included in the package as knowledge or interest in programming is not expected in the luthier's training. A screen capture of the GUI is shown in \autoref{fig:gui}. The design of guitarsounds differs from existing packages in its ability to be both used alone to produce decent figures with a minimal number of lines of code and as a tool in a python data visualization stack where the sound specific needs can be handled by guitarsounds. More over, the GUI acts as a different way of using the same tool in the same environment (Jupyter Notebook) as to allow for institutional uniformity in teaching and experimentation. There exists overlap between guitarsounds and the librosa [@mMcFee:2015] python package for music analysis, however librosa is not a dependency of guitarsounds and the latter is more focused on feature extraction for machine learning application and lacks features tailored for harmonic sound analysis and  integrated comparison of sounds. Librosa is also aimed at an audience with a more developed knowledge of python programming.
+Guitarsounds was written to meet the needs of the Bruant lutherie school, more precisely as a tool to visualize and compare the sounds of different guitar designs based on arbitrary sound features. The guitarsounds API serves as a introduction to programming for data analysis for the school's students and the GUI allows visualizing sound features even for users with minimal affinity with computers, as such is the case for some students. More over, the guitarsounds package is used in the teaching activities as a tool to visualize the physical phenomenons involved in the sound produced by guitars, such as the hemlotz cavity frequency of the instrument. The GUI was included in the package as knowledge or interest in programming is not expected in the luthier's training. A screen capture of the GUI is shown in \autoref{fig:gui}. The design of guitarsounds differs from existing packages in its ability to be both used alone to produce decent figures with a minimal number of lines of code and as a tool in a python data visualization stack where the sound specific needs can be handled by guitarsounds. More over, the GUI acts as a different way of using the same tool in the same environment (Jupyter Notebook) as to allow for institutional uniformity in teaching and experimentation. There exists overlap between guitarsounds and the librosa [@Librosa] python package for music analysis, however librosa is not a dependency of guitarsounds and the latter is more focused on feature extraction for machine learning application and lacks features tailored for harmonic sound analysis and  integrated comparison of sounds. Librosa is also aimed at an audience with a more developed knowledge of python programming.
 
 ![Graphical user interface in the Jupyter Notebook environment.\label{fig:gui}](figuregui.png){width=50%}
 
