+++ paper/paper.bib
@@ -7,3 +7,16 @@
 	       pages = {18--24},
 	       }
 
+@article{sumi_classical_2008,
+	title = {Classical guitar top board design by finite element method modal analysis based on acoustic measurements of guitars of different quality},
+	volume = {29},
+	issn = {1347-5177, 1346-3969},
+	url = {http://joi.jlc.jst.go.jp/JST.JSTAGE/ast/29.381?from=CrossRef},
+	doi = {10.1250/ast.29.381},
+	language = {en},
+	number = {6},
+	journal = {Acoustical Science and Technology},
+	author = {Sumi, Takeshi and Ono, Teruaki},
+	year = {2008},
+	pages = {381--383},
+}
