### 2022-09-16 #4769 @olivecha
Hi daniel,

Re 1 : There was a problem with the way the references citations were included, it is fixed now.

Re 2 : A description of the research used was added in the "statement of need"  section
