### 2022-11-14 olivecha/guitarsounds#33 @cwitkowitz
Loading sound file crashes GUI
I receive the following errors when attempting to load a sound file through each button of the GUI:

**Single Sound**
```
File ~/Desktop/guitarsounds/guitarsounds/interface.py:783, in guitarGUI.define_sound_names(self)
    779     # Case for a single sound analysis
    780     if self.analysis == 'Single':
    781
    782         # get the filenames
--> 783         self.file_names = [ky for ky in self.single_file_selector.value.keys()]

AttributeError: 'tuple' object has no attribute 'keys'
```

**Dual Sounds**
```
File ~/Desktop/guitarsounds/guitarsounds/interface.py:819, in guitarGUI.define_sound_names(self)
    817     if self.analysis == 'Dual':
    818         # get the file names
--> 819         name1 = [ky for ky in self.dual_file_selector_1.value.keys()][0]
    820         name2 = [ky for ky in self.dual_file_selector_2.value.keys()][0]

AttributeError: 'tuple' object has no attribute 'keys'
```

**Multiple Sounds**
```
File ~/Desktop/guitarsounds/guitarsounds/interface.py:824, in guitarGUI.define_sound_names(self)
    823 elif self.analysis == 'Multiple':
--> 824     self.file_names = [ky for ky in self.mult_file_selector.value.keys()]

AttributeError: 'tuple' object has no attribute 'keys'
```

Each results in a crash.
