### 2022-11-16 #4878 @cwitkowitz
> wave Hi all - just checking in for any updates on this one?

I am currently in the process of completing my review. So far, I have read the software paper, scanned through the repository, and opened up several issues. Issue olivecha/guitarsounds#33 is quite significant, as it prevents experimentation with the GUI. I still need to go through and verify the API functionality and skim through the documentation. After that, I will organize my thoughts/comments and finalize my initial review. Hopefully I can finish up this weekend.
