### 2022-11-22 olivecha/guitarsounds#44 @cwitkowitz
Comments on Paper
### General
- There is room for improvement with respect to the language and phrasing used throughout the paper. For instance, I had to read the second sentence of the paper a few times until I realized that the final part (split by comma) refers to "transient harmonic sounds". Some sentences read a bit clunky, e.g. sentences using "allows" or "allowing".
- Correct me if I'm wrong, but there is nothing inherently specific to guitar in this framework (besides maybe the assumption of strong and quick onsets), right? If so, it would be good to highlight that the framework can be used for _any_ harmonic sounds.
- When referencing python packages, such as guitarsounds, matplotlib, or librosa, I think it would be better to treat these as code snippets, i.e. `guitarsounds`, `matplotlib`, or `librosa`. This will make them pop out in the paper more and will draw attention to the fact that they are packages. On a similar note, I'm not sure if guitarsounds should be capitalized, since the name of the package is not capitalized.
- I think some things in this paper may be confusing to a non-specialist audience, such as some mentioned features (e.g. envelope) or luthier terminology (e.g. soundboard).
- I don’t think the software paper does full justice to the package at this point. There are a lot of distinctions, i.e. signal features, sounds, and sound packs, as well as the corresponding visualizations of each, which are not really highlighted.

Here are some spelling considerations:
- "envelop" -> "envelope"
- "hemlotz" -> "Helmholtz"
-  “a” and “an” are frequently mixed up
- “Bruant” or “Bruand” - is this a difference between the English and French title?
- Capitalization of "conditioning"
- "WAVE" vs. "wave" inconsistency
- "More over" -> "Moreover"
- Capitalize "Python"?
- "tailored for" -> "tailored to"

### Summary
- I'm not sure if it’s appropriate to claim that `guitarsounds` implements the FFT. It just provides a wrapper around NumPy.
- Run-on sentence: "The guitarsounds package is divided ...". Add a comma or split.
- Are loading/saving/normalization really main features? They seem relatively trivial.
- I think it is more appropriate to refer to frequency "bands" rather than "bins". In my experience, a frequency bin corresponds to a single frequency, whereas a frequency band corresponds to a group of such adjacent bins.
- Redundant mention of Fourier transform peaks in the main features list.
- It seems a lot of the visualization capabilities are left out of the main features list. This may lead readers to missing the full picture with respect to what is available in the package.
- I don't understand the final item in the main features list.

### Example
- The description of the example is quite vague and, at a surface level, does not appear to correspond to the code snippet provided. It could be improved by making the details of the conditioning more explicit and mentioning that this is happening under the hood in `guitarsounds.SoundPack.__init__`.
- See the relevant comments I made about this type of plot in issues #39 and #40. Also note that this plot, while generated from the same code, is different from what is displayed in `README.md`. Furthermore, in the paper, the difference between "peaks" and "diff peaks" goes unexplained.
- After the example it would be good to mention that there are other visualization function sets available as well, i.e. with the `Sound` and `Plot` classes.
- It may be better to stick with a single example throughout the paper for a more holistic motivation of the package. For instance, in the summary, you exemplify the log-time envelope of a sound. You could described what it is, why it is useful, and visualize that instead with `guitarsounds` here.

### Statement of Need
- Can the previous works (i.e. innovative numerical prototyping method) be cited?
- How is `guitarsounds` used to modify sounds?
- Awkward wording: "slope of a linear regression".
- Awkward wording: "as such is the case".
- Plural of "phenomenon" is "phenomena".
- Awkward wording: "GUI acts as a different way".

### References
- I think the Numpy and SciPy Python packages should probably be cited, since they are heavily utilized in the `guitarsounds` framework. The paper should also make note of this utilization.
