### 2022-11-25 olivecha/guitarsounds#45 @ebezzam
Bad parsing of "all" for "Frequency bin plot"
When using the interface for the "Frequency bin plot" for "all", I get the following error:

```
File ~/GITHUB/guitarsounds/guitarsounds/interface.py:589, in guitarGUI.on_go_button_clicked(self, b)
    587 plt.figure(figsize=(8, 6))
    588 # Call the method
--> 589 self.analysis_tuple[0](self.sounds, bins=[self.analysis_tuple[1]])
    591 # Define the title according to the chosen bin
    592 if self.analysis_tuple[1] == 'all':

File ~/GITHUB/guitarsounds/guitarsounds/analysis.py:959, in Sound.plot_freq_bins(self, bins)
    956         bins = [bins]
    958 for key in bins:
--> 959     lab = key + ' : ' + str(int(self.bins[key].range[0])) + ' - ' + str(int(self.bins[key].range[1])) + ' Hz'
    960     self.bins[key].old_plot('log envelop', label=lab)
    962 plt.xscale('log')

KeyError: 'all'
```

I think because there's a bad parsing at https://github.com/olivecha/guitarsounds/blob/edc3c05e0cf508c10956f611157070cb3bbac0a3/guitarsounds/analysis.py#L952

I think you may be passing `["all"]` so your check isn't able to identify that the user is request all plots as it is not a string.
