### 2022-11-25 olivecha/guitarsounds#33 @ebezzam
I didn't have any issue when loading sound(s) through each of the buttons. I'm wondering what could cause a difference. You mind sharing your Python version and package versions?
