### 2022-11-25 olivecha/guitarsounds#46 @ebezzam
README example
In the example code block https://github.com/olivecha/guitarsounds/blob/edc3c05e0cf508c10956f611157070cb3bbac0a3/README.md?plain=1#L57

it may be obvious, but mention the import statement as well `from guitarsounds import Sound`. Because there could be sub-modules.
