### 2022-11-25 olivecha/guitarsounds#46 @ebezzam
Same for `Soundpack` later on.
