### 2022-11-25 olivecha/guitarsounds#46 @ebezzam
And for the remaining code examples, please provide the code in code-blocks (rather than screenshots) so that it is easy to copy-and-paste for users. And you can still put screenshots of the resulting plots.
