### 2022-11-25 olivecha/guitarsounds#46 @ebezzam
Moreover, you had a [tutorial notebook](https://github.com/olivecha/guitarsounds/blob/main/docs/API_Tutorial.ipynb)! I would recommend mentioning + linking that in the README.
