### 2022-11-25 olivecha/guitarsounds#47 @ebezzam
No `timbre` method
When running the demo notebook, I get the following error when running `mysound.plot.timbre(label='regular')`

```
Input In [30], in <cell line: 5>()
      1 # # Timbral attributes not loud
      2 # mysound.plot.timbre(label='regular')
      3
      4 # Timbral attributes loud
----> 5 mysound.signal.normalize().plot.timbre(label='louder')
      7 # add labels to the legend
      8 guitarsounds.plt.legend()

AttributeError: 'Plot' object has no attribute 'timbre'
```

It seems like there is no `timbre` method for the [`Plot`](https://github.com/olivecha/guitarsounds/blob/edc3c05e0cf508c10956f611157070cb3bbac0a3/guitarsounds/analysis.py#L1561) object.
