### 2022-11-25 olivecha/guitarsounds#44 @ebezzam
To add this very exhaustive list (thanks @cwitkowitz):

- The sentence at Lines 12 - 16 could be broken up so that it's clearer what the two components are, e.g. (1) and (2) just before the components could make it clearer.
- API stands for Application Programming Interface (13-14)
- some wording that I think are missing "-": well-defined (10), object-oriented (13), time-dependent (28), easy-to-use (35)
- Finish each bullet with "." or ";" (until the last with ".").
- Line 42: "two or more sounds."

And to reiterate two of @cwitkowitz comments:
- I think you could broaden the scope and visibility by not renaming the package, as the techniques are not limited to guitar sounds.
- Citing NumPy and Matplotlib
