### 2022-11-25 #4878 @ebezzam
@bmcfee I've finished my review. I also recommend revisions before acceptance. Globally I think it's a useful package that provides (1) signal processing techniques to conveniently analyze and compare different aspects of harmonic signals and (2)  a solution that is accessible for non-programming saavy people.

If possible, I would consider re-naming the package as I believe these techniques are not limited to just guitar sounds. This could broaden the interest for the package.

The GUI worked for me with [one feature](https://github.com/olivecha/guitarsounds/issues/45) that crashed the GUI.

The demo notebook was useful, and I would definitely mention it in the README. For example, at the very top so that people who end up on the GitHub page can directly see working code. The `timbre` functionality didn't work for me, as I've mentioned [here](https://github.com/olivecha/guitarsounds/issues/47).

[Here](https://github.com/olivecha/guitarsounds/issues/46) I've mentioned other things I would change in the README. I'm torn on keeping all the Anaconda installation instruction in the README. I get that it is to cater for people who may not be familiar with Python programming, but it takes up a lot of space and on first glance detracts from the main purpose of the package.

Re Documentation
- A small note on how to run the test would be useful, e.g.
```
cd tests/
pytest
```
- A `CONTRIBUTING.md` file (and a optionally a CHANGELOG) would be useful to provide clear community guidelines.

Re Software Paper
- I added some comments to the issue started by @cwitkowitz: https://github.com/olivecha/guitarsounds/issues/44
