### 2022-12-12 #4878 @olivecha
Hi, yes I looked at the issues, sorry I was quite busy in the previous weeks. I will try and get them solved by the end of 2022.

Thank you very much for the exhaustive review, I really appreciate the time taken here !
