### 2022-12-23 3b98e1c
Update API tutorial to match changes in source issue #47
