### 2022-12-23 olivecha/guitarsounds#47 @olivecha
Hi,
The `timbre` method was removed at one point as it was not particularly useful in evaluating sounds in the lutherie domain (not a lot meaningful comparisons could be obtained because the underlying timbre tool was meant to compare sounds with stronger differences).

commit 3b98e1c6febbac585fd91aa20128a1952365387f solved the issue

Thanks,

Olivier
