### 2022-12-23 cb383fb
Added code snippets to allow copy and paste
