### 2022-12-23 b44b419
Add API tutorial in  README.md for issue #46
