### 2022-12-23 olivecha/guitarsounds#46 @olivecha
Hi,

I did the prescribed changes to the README.md (commit b44b41911d25d9212e0f661aef026837ea0a5b69 and others before)

Good idead to add the API tutorial in the README.md
