### 2022-12-23 240b6a1
modify Sound.plot_freq_bin parsing issue #45
