### 2022-12-23 olivecha/guitarsounds#45 @olivecha
Hi,

Indeed there was a parsing problem from the interface to the analysis module. The issue was fixed by commit 240b6a1ccf076e88397112467482bf8d2c848b70.
