### 2023-01-09 923dd6e
Fix for issue #42
