+++ guitarsounds/interface.py
@@ -10,10 +10,16 @@ import numpy as np
 
 
 def generate_error_widget(text):
+    """
+    Creates a ipywidget error message from a string
+    :param text: python str of the error message
+    :return: widgets.HTLM object corresponding to the error message
+    """
     return widgets.HTML('<p style="color:#CC4123;">' + text + '</p>')
 
 
 class guitarGUI(object):
+    """ Main Graphical user interface class """
     # Output layout
     out_layout = {'border': '1px solid black'}
 
@@ -21,13 +27,20 @@ class guitarGUI(object):
     box_layout = widgets.Layout(align_items='stretch', flex_flow='line', width='50%')
 
     # Fundamental input style
-    fundamental_style = style = {'description_width': 'initial'}
+    style = {'description_width': 'initial'}
 
     # Attribute for output layout
     output = widgets.Output(layout={'border': '1px solid black'})
 
     # List of plot methods
-    plot_methods = [Plot.signal, Plot.envelop, Plot.log_envelop, Plot.fft, Plot.fft_hist, Plot.peaks, Plot.peak_damping,
+    plot_methods = [Plot.signal, 
+                    Plot.envelop, 
+                    Plot.log_envelop, 
+                    Plot.fft, 
+                    Plot.fft_hist, 
+                    Plot.peaks, 
+                    Plot.peak_damping,
+                    Plot.integral,
                     Plot.time_damping, ]
     bin_ticks_methods = [Plot.fft, Plot.fft_hist, Plot.peaks, Plot.peak_damping, ]
 
@@ -57,6 +70,7 @@ class guitarGUI(object):
                ('Peaks plot', Plot.peaks),
                ('Peak damping plot', Plot.peak_damping),
                ('Time damping plot', Plot.time_damping),
+               ('Cumulative integral plot', Plot.integral),
                ('Frequency damping values', Sound.peak_damping), ]
 
     drop_down_style = {'description_width': '150px'}
@@ -127,7 +141,6 @@ class guitarGUI(object):
                ('Fourier Transform Peaks', 'peaks'),
                ('Peak Damping', 'peak damping'),
                ('Time Damping', 'time damping'),
-               ('Timbre Attributes', 'timbre'),
                ('Cumulative integral', 'integral'), ]
 
     plot_drop_down = widgets.Dropdown(options=options, value='signal', style=drop_down_style,
@@ -137,12 +150,14 @@ class guitarGUI(object):
 
     def __init__(self):
         """
-        Here We display the three file choosing buttons matched with the
+        Constructor for the Graphical User Interface class
+
+        Upon instanciation of the class, we display the three file choosing buttons matched with the
         three types of analyses when one is clicked the user is prompted
         to choose files.
 
         When files are chosen  the user press the 'Ok' Button and the
-        Program advances to defining names see `.on_ok_button_clicked_1`.
+        Interface  advances to defining names see `.on_ok_button_clicked_1`.
         """
 
         # __ Buttons __
@@ -207,7 +222,7 @@ class guitarGUI(object):
         self.button2.on_click(self.on_dual_button_clicked)
         self.button3.on_click(self.on_multiple_button_clicked)
         self.ok_button.on_click(self.on_ok_button_clicked_1)
-        self.disable_file_selection(False)
+        self.change_file_selection_state(False)
 
         # display the buttons
         display(self.button_box)
@@ -220,13 +235,14 @@ class guitarGUI(object):
         """
         Displays the single file selector, allowing the user to choose
         one file.
+        :param: b the ipywidget button object for which this method is executed when it is clicked
         """
         if b is not None:
             pass
         clear_output(wait=True)
 
         output = widgets.Output(layout={'border': '1px solid black'})
-        self.disable_file_selection(True)
[…]
