### 2023-01-10 ecf672d
refractor utils.py (issue #41)
