+++ guitarsounds/utils.py
@@ -1,24 +1,38 @@
+import struct
+import wave
 import numpy as np
 import scipy.optimize
 import scipy.integrate
 from scipy.interpolate import interp1d
 import guitarsounds
 from guitarsounds.parameters import sound_parameters
-import wave
-import struct
 
 # instantiation of the trim time interpolator
 sp = sound_parameters()
-freq_dict = {'E2': 82.41, 'A2': 110.0, 'D3': 146.83, 'G3': 196.0, 'B3': 246.94, 'E4': 329.63}
-trim_dict = {'E2': sp.trim.E2.value, 'A2': sp.trim.A2.value, 'D3': sp.trim.D3.value, 'G3': sp.trim.G3.value,
-             'B3': sp.trim.B3.value, 'E4': sp.trim.E4.value}
-freq2trim = interp1d(list(freq_dict.values()), list(trim_dict.values()), fill_value='extrapolate')
+freq_dict = {'E2': 82.41,
+             'A2': 110.0,
+             'D3': 146.83,
+             'G3': 196.0,
+             'B3': 246.94,
+             'E4': 329.63}
 
+trim_dict = {'E2': sp.trim.E2.value,
+             'A2': sp.trim.A2.value,
+             'D3': sp.trim.D3.value,
+             'G3': sp.trim.G3.value,
+             'B3': sp.trim.B3.value,
+             'E4': sp.trim.E4.value}
+
+freq2trim = interp1d(list(freq_dict.values()),
+                     list(trim_dict.values()),
+                     fill_value='extrapolate')
 
 def nth_order_polynomial_residual(A, n, x, y):
     """
-    Function computing the residual from a nth order polynomial, with x and y vectors
-    :return:
+    Function computing the residual from a nth order
+    polynomial, with x and y vectors
+    :return: the evaluated residual between a polynomial and the polynomial
+    with coefficients A
     """
     y_model = 0
     for i in np.arange(n):
@@ -28,7 +42,8 @@ def nth_order_polynomial_residual(A, n, x, y):
 
 def nth_order_polynomial_fit(n, x, y):
     """
-    Function creating a function of a fitted nth order polynomial to a set of x and y data
+    Function creating a function of a fitted nth order
+    polynomial to a set of x and y data
     :param n: polynomial order
     :param x: x values
     :param y: y values
@@ -38,7 +53,9 @@ def nth_order_polynomial_fit(n, x, y):
     while n > len(x):
         n -= 1
     guess = np.ones(n)
-    result = scipy.optimize.least_squares(nth_order_polynomial_residual, guess, args=(n, x, y))
+    result = scipy.optimize.least_squares(nth_order_polynomial_residual,
+                                          guess,
+                                          args=(n, x, y))
     A = result.x
 
     def polynomial_function(x_value):
@@ -53,8 +70,13 @@ def nth_order_polynomial_fit(n, x, y):
 def octave_values(fraction, min_freq=10, max_freq=20200):
     """
     Compute octave fraction bin center values from min_freq to max_freq.
-    `soundcomp.octave_values(3)` returns the bins corresponding to 1/3 octave values
+    `soundcomp.octave_values(3)` returns the bins corresponding to 1/3 octave
+    values
     from 10 Hz to 20200 Hz.
+    :param fraction: octave fraction value (1/3, 1, 1/2, ...)
+    :param min_freq: minimum frequency in the computed octave fracitons
+    :param max_freq: maximum frequency considered in the octave fractions
+    :return: the value in frequencies corresponding to the octave bins
     """
     # Compute the octave bins
     f = 1000.  # Reference Frequency
@@ -76,6 +98,10 @@ def octave_histogram(fraction, **kwargs):
     """
[…]
