+++ guitarsounds/analysis.py
@@ -1007,7 +1007,13 @@ class Signal(object):
     """
 
     def __init__(self, signal, sr, SoundParam, freq_range=None):
-        """ Create a Signal class from a vector of samples and a sample rate"""
+        """ 
+        Create a Signal class from a vector of samples and a sample rate
+        :param signal: vector containing the signal samples
+        :param sr: sample rate of the signal (Hz)
+        :param SoundParam: Sound Parameter instance to use with the signal
+        :para freq_range: Frequency range to use with the signal
+        """
         self.SP = SoundParam
         self.onset = None
         self.signal = signal
@@ -1021,9 +1027,12 @@ class Signal(object):
     def time(self):
         """
         Returns the time vector associated to the signal
-        :return: numpy array corresponding to the time values of the signal samples in seconds
+        :return: numpy array corresponding to the time values
+        of the signal samples in seconds
         """
-        return np.linspace(0, len(self.signal) * (1 / self.sr), len(self.signal))
+        return np.linspace(0,
+                           len(self.signal) * (1 / self.sr),
+                           len(self.signal))
 
     def listen(self):
         """
@@ -1046,9 +1055,7 @@ class Signal(object):
 
         Calls the function corresponding to Plot.kind()
         See help(guitarsounds.analysis.Plot) for info on the different plots
-        (not tested)
         """
-
         self.plot.method_dict[kind](**kwargs)
 
     def fft(self):
@@ -1068,75 +1075,16 @@ class Signal(object):
         SC = np.sum(self.fft() * self.fft_frequencies()) / np.sum(self.fft())
         return SC
 
-    def peaks_old(self, max_freq=None, height=False, result=False):
-        """
-        Computes the harmonic peaks indexes from the FFT of the signal
-        :param max_freq: Supply a max frequency value overriding the one in guitarsounds_parameters
-        :param height: if True the height threshold is returned to be used in the 'peaks' plot
-        :param result: if True the Scipy peak finding results dictionary is returned
-        :return: peak indexes
-        """
-        # Replace None by the default value
-        if max_freq is None:
-            max_freq = self.SP.general.fft_range.value
-
-        # Get the fft and fft frequencies from the signal
-        fft, fft_freq = self.fft(), self.fft_frequencies()
-
-        # Find the max index
-        max_index = np.where(fft_freq >= max_freq)[0][0]
-
-        # Find an approximation of the distance between peaks, this only works for harmonic signals
-        peak_distance = np.argmax(fft) // 2
-
-        # Maximum of the signal in a small region on both ends
-        fft_max_start = np.max(fft[:peak_distance])
-        fft_max_end = np.max(fft[max_index - peak_distance:max_index])
-
-        # Build the curve below the peaks but above the noise
-        exponents = np.linspace(np.log10(fft_max_start), np.log10(fft_max_end), max_index)
-        intersect = 10 ** exponents[peak_distance]
-        diff_start = fft_max_start - intersect  # offset by a small distance so that the first max is not a peak
-        min_height = 10 ** np.linspace(np.log10(fft_max_start + diff_start), np.log10(fft_max_end), max_index)
-
-        first_peak_indexes, _ = sig.find_peaks(fft[:max_index], height=min_height, distance=peak_distance)
-
-        number_of_peaks = len(first_peak_indexes)
-        if number_of_peaks > 0:
-            average_len = int(max_index / number_of_peaks) * 3
-        else:
-            average_len = int(max_index / 3)
-
-        if average_len % 2 == 0:
-            average_len += 1
-
-        average_fft = sig.savgol_filter(fft[:max_index], average_len, 1, mode='mirror') * 1.9
-        min_freq_index = np.where(fft_freq >= 70)[0][0]
-        average_fft[:min_freq_index] = 1
-
[…]
