### 2023-01-11 9d14e51
add test for utils.py issue #43
