### 2023-01-11 3e11313
added more utils.py tests #43
