### 2023-01-11 8a561d5
add function call in test #43
