+++ tests/test_SoundPack.py
@@ -1,4 +1,6 @@
 import unittest
+from contextlib import redirect_stdout
+from io import StringIO
 import guitarsounds
 from guitarsounds import SoundPack
 import helpers_tests
@@ -82,10 +84,137 @@ class MyTestCase(unittest.TestCase):
             plt.close(plt.gcf())            
     
     def test_SoundPack_listen(self):
+        """ Test the SoundPack.listen method """
         files = [helpers_tests.get_rnd_audio_file() for _ in range(4)]
         sp = SoundPack(files)
         sp.listen()
         self.assertTrue(True)
+
+    def test_soundpack_freq_bin_plot(self):
+        """ test the SoundPack.freq_bin_plot method """
+        # Create a SoundPack instance
+        files = [helpers_tests.get_rnd_audio_file() for _ in range(3)]
+        sp = SoundPack(files)
+        # Test branch with all frequency bins
+        sp.freq_bin_plot(f_bin='all')
+        # test that all the frequency bins are plotted
+        self.assertTrue(len(plt.gcf().axes) == 6)
+        plt.close(plt.gcf())
+        # Test branch with a specific frequency bin
+        sp.freq_bin_plot(f_bin='mid')
+        self.assertTrue(len(plt.gcf().axes) == 1)
+
+    def test_soundpack_combine_envelop(self):
+        """ the the combine_envelop method of SoundPack """
+        files = [helpers_tests.get_rnd_audio_file() for i in range(3)]
+        sp = SoundPack(files)
+        # Test signal plot branch
+        sp.combine_envelop(kind='signal')
+        # the plot is supposed to have lines
+        self.assertTrue(len(plt.gcf().axes[0].lines) >= 0)
+        plt.close(plt.gcf())
+        # test the branch with a frequency bin and only showing the combined
+        # curve
+        sp.combine_envelop(kind='bass', show_sounds=False, show_rejects=False)
+        self.assertTrue(len(plt.gcf().axes[0].lines) == 1)
+
+    def test_soundpack_fundamentals(self):
+        """ test the fundamentals method of SoundPack """
+        # create a SoundPack instance
+        files = [helpers_tests.get_rnd_audio_file() for i in range(3)]
+        sp = SoundPack(files)
+        # redirect_stdout
+        output = StringIO()
+        with redirect_stdout(output):
+            sp.fundamentals()
+        self.assertTrue(len(output.getvalue()) > 0)
+
+    def test_soundpack_integral_plot(self):
+        """ test the SoundPack integral_plot method """
+        # Create a soundpack
+        files = [helpers_tests.get_rnd_audio_file() for i in range(3)]
+        sp = SoundPack(files)
+        # Test branch with all frequency bins
+        sp.integral_plot(f_bin='all')
+        self.assertTrue(len(plt.gcf().axes) == 6)
+        # Test branch with a specific frequency bin
+        sp.integral_plot(f_bin='highmid')
+        self.assertTrue(len(plt.gcf().axes) == 1)
+
+    def test_soundpack_bin_power_table(self):
+        """ Test the SoundPack bin_power_table method """
+        # create a SoundPack instance
+        files = [helpers_tests.get_rnd_audio_file() for i in range(3)]
+        sp = SoundPack(files)
+        # redirect_stdout
+        output = StringIO()
+        with redirect_stdout(output):
+            sp.bin_power_table()
+        self.assertTrue(len(output.getvalue()) > 0)
+
+    def test_soundpack_bin_power_hist(self):
+        """ Test the bin_power_hist method of SoundPack """
+        # create a SoundPack instance
+        files = [helpers_tests.get_rnd_audio_file() for i in range(3)]
+        sp = SoundPack(files)
[…]
