### 2023-01-11 1a7ecf9
added tests for interface.py #43
