### 2023-01-11 f5e8518
debug signal.log_envelop test
