### 2023-01-11 8c3b7ca
remove combine_envelop method of SoundPack
