+++ guitarsounds/analysis.py
@@ -306,90 +306,6 @@ class SoundPack(object):
         else:
             print('invalid frequency bin')
 
-    def combine_envelop(self, kind='signal', difference_factor=1, show_sounds=True, show_rejects=True, **kwargs):
-        """
-        __ Multiple SoundPack Method __
-        Combines the envelopes of the Sounds contained in the SoundPack, Sounds having a too large difference factor
-        from the average are rejected.
-        :param kind: which signal to use from :
-        'signal', 'bass', 'mid', 'highmid', 'uppermid', 'presence', 'brillance'
-        :param difference_factor: threshold to reject a sound from the combination, can be adjusted to reject
-        or include more sounds.
-        :param show_sounds: If True all the included Sounds are shown on the plot
-        :param show_rejects: If True all the rejected Sounds are shown on the plot
-        :param kwargs: Key word arguments to pass to the envelope plot.
-        :return: None
-        """
-        sounds = self.sounds
-        sample_number = np.min([len(s1.signal.log_envelop()[0]) for s1 in sounds])
-
-        if kind == 'signal':
-            log_envelops = np.stack([s1.signal.normalize().log_envelop()[0][:sample_number] for s1 in sounds])
-        elif kind in SP.bins.__dict__.keys():
-            log_envelops = np.stack([s1.bins[kind].normalize().log_envelop()[0][:sample_number] for s1 in sounds])
-        else:
-            raise ValueError
-
-        average_log_envelop = np.mean(log_envelops, axis=0)
-        means = np.tile(average_log_envelop, (len(sounds), 1))
-        diffs = np.sum(np.abs(means - log_envelops), axis=1)
-        diff = np.mean(diffs) * difference_factor
-
-        good_sounds = np.array(sounds)[diffs < diff]
-        rejected_sounds = np.array(sounds)[diffs > diff]
-        average_log_envelop = np.mean(log_envelops[diffs < diff], axis=0)
-        norm_factors = np.array([s1.signal.normalize().norm_factor for s1 in good_sounds])
-
-        if kind == 'signal':
-            if show_sounds:
-                for s1 in good_sounds[:-1]:
-                    s1.signal.normalize().old_plot(kind='log envelop', alpha=0.2, color='k')
-                sounds[-1].signal.normalize().old_plot(kind='log envelop', alpha=0.2, color='k', label='sounds')
-
-            if show_rejects:
-                if len(rejected_sounds) > 1:
-                    for s1 in rejected_sounds[:-1]:
-                        s1.signal.normalize().old_plot(kind='log envelop', alpha=0.3, color='r')
-                    rejected_sounds[-1].signal.normalize().old_plot(kind='log envelop', alpha=0.3, color='r',
-                                                                    label='rejected sounds')
-                if len(rejected_sounds) == 1:
-                    rejected_sounds[0].signal.normalize().plot(kind='log envelop', alpha=0.3, color='r',
-                                                               label='rejected sounds')
-            if len(good_sounds) > 0:
-                if 'label' in kwargs.keys():
-                    plt.plot(good_sounds[0].signal.log_envelop()[1][:len(average_log_envelop)], average_log_envelop,
-                             **kwargs)
-                else:
-                    plt.plot(good_sounds[0].signal.log_envelop()[1][:len(average_log_envelop)], average_log_envelop,
-                             label='average', color='k', **kwargs)
-
-        else:
-            if show_sounds:
-                for s1 in good_sounds[:-1]:
-                    s1.bins[kind].normalize().old_plot(kind='log envelop', alpha=0.2, color='k')
-                sounds[-1].bins[kind].normalize().old_plot(kind='log envelop', alpha=0.2, color='k', label='sounds')
-
[…]
