### 2023-01-16 a5f4f1c
Work on paper #44
