### 2023-01-16 olivecha/guitarsounds#44 @olivecha
Thanks for the comments on the paper

I've done the following changes :
- Worked on general phrasing issues.
- Specified the possibility of using the package for any harmonic sound.
- Write package names as code snippets
- Further defined specific terminology
- Highlight how the different features are implemented in the package classes structures
- Added citation for `numpy` and `scipy` Python packages
- Some spelling fixes
