+++ paper/paper.md
@@ -1,4 +1,4 @@
----
+--
 title: 'Guitarsounds : A Python package to visualize harmonic sounds for musical instrument testing'
 tags:
   - Python
@@ -22,11 +22,12 @@ bibliography: paper.bib
 
 # Summary
 
-The `guitarsounds` python package can be used to rapidly visualize relevant features of harmonic sounds and is mainly developed to compare musical instrument design iterations. 
+The `guitarsounds` Python package can be used to rapidly visualize relevant features of harmonic sounds and is mainly developed to compare musical instrument design iterations. 
 `guitarsounds` is wrapped around the implementation of common signal processing features. 
 The implementations used in the `guitarsounds` package are manly taken from `numpy` [@harris_array_2020] and `scipy` [@virtanen_scipy_2020]. 
 The signal processing features are augmented to perform the analysis of transient harmonic sounds.
-Such sounds are defined as having a defined onset and a frequency-amplitude distribution clearly concentrated around defined partials. Consequently, while the package is named `guitarsounds`, its analysis framework can be used with any harmonic sound, such as those produced by a piano or a percussion instrument. 
+Such sounds are defined as having a defined onset, and a frequency-amplitude distribution clearly concentrated around defined partials. 
+Consequently, while the package is named `guitarsounds`, its analysis framework can be used with any harmonic sound, such as those produced by a piano, or a percussion instrument. 
 
 The guitarsounds package is divided in two main components. 
 First, the package is developped around a convenient object oriented Application Programming Interface (API) which can be used to extract features from sounds and visualize them according to the user's needs. 
@@ -37,13 +38,13 @@ The main features of guitarsounds are:
 
 - Automate the loading, conditioning and normalization of multiple sound files to meaningfully compare their features.
 - Visualize sounds features relevant to musical instrument design, such as: 
-    - Linear and logarithmic time envelop
+    - Linear and logarithmic time envelope
     - Octave bins Fourier transform
     - Time dependent damping
 - Divide sounds in frequency bins to analyze variations in temporal behaviour for different frequency ranges.
-- Extract the Fourier transform peaks of an harmonic signal using a custom peak finding algorithm.
+- Extract the Fourier transform peaks of a harmonic signal using a custom peak finding algorithm.
 - Extract numerical values for certain features such as the Helmholtz cavity frequency of a guitar.
-- Provide a easy to use signal processing API to compute new features meeting to specific needs by providing acess to lower level features and handling the differences between sound files, such as the file sample rate.
+- Provide an easy-to-use signal processing API to compute new features meeting to specific needs by providing acess to lower level features and handling the differences between sound files, such as the file sample rate.
 
 Specifically, the API provides 5 classes nested together : `SoundPack`, `Sound`, `Signal`, `Plot`. 
 Starting from the innermost class, the `Plot` class handles the low level plotting of specific features, such as plotting the FFT for a single sound file.
@@ -60,10 +61,10 @@ As an example, the method plotting the FFT of two sounds in a mirror configurati
 
 ###### TODOOOO : rewrite example using the log-time envelop
 
-As an example, the log-time envelop, which describes the amplitude of a sound with a higher refinement in time at the start of the sound, can be computed for a sound filtered within a specific frequency bandwidth. 
-By plotting such a feature for the same note played on two instruments, the dynamic response of the instruments can be compared for a specific frequency range.
+As an example, the log-time envelope, which describes the amplitude of a sound with a higher refinement in time at the start of the sound, can be computed for a sound filtered within a specific frequency bandwidth. 
+By plotting this feature for the same note played on two instruments, the dynamic response of the instruments can be compared for a specific frequency range.
 
[…]
