### 2023-01-17 869dda2
Improve paper #44
