+++ paper/paper.md
@@ -39,9 +39,9 @@ The main features of guitarsounds are:
 - Automate the loading, conditioning and normalization of multiple sound files to meaningfully compare their features.
 - Visualize sounds features relevant to musical instrument design, such as: 
     - Linear and logarithmic time envelope
-    - Octave bins Fourier transform
+    - Octave bands Fourier transform
     - Time dependent damping
-- Divide sounds in frequency bins to analyze variations in temporal behaviour for different frequency ranges.
+- Divide sounds in frequency bands to analyze variations in temporal behaviour for different frequency ranges.
 - Extract the Fourier transform peaks of a harmonic signal using a custom peak finding algorithm.
 - Extract numerical values for certain features such as the Helmholtz cavity frequency of a guitar.
 - Provide an easy-to-use signal processing API to compute new features meeting to specific needs by providing acess to lower level features and handling the differences between sound files, such as the file sample rate.
@@ -54,46 +54,55 @@ An instance of the `Plot` class is constructed for each `Signal` class instance 
 The `Signal` class provides all the features relying only on the data of a single sound signal as class methods. 
 The `Sound` class is used to store all the information corresponding to a single sound file. 
 When a `.wav` file is read, all the processing is handled by the `Sound` class, such as truncating, filtering or normalizing the sound signal. 
-The `Sound` class provides the features relying on more than one `Signal` instance, but still using the information from a single sound file, such as the power distribution of a sound across different frequency bins. 
+The `Sound` class provides the features relying on more than one `Signal` instance, but still using the information from a single sound file, such as the power distribution of a sound across different frequency bands. 
 Finally, the `SoundPack` class is constructed from multiple `Sound` class instances and provides the features used to compare the data between different sound files. 
 The `SoundPack` methods are divided between methods developed to compare two sounds and methods developed to compare an arbitrary amount of sounds. 
 As an example, the method plotting the FFT of two sounds in a mirror configuration can only be called if the `SoundPack` was constructed using exactly two sounds, whereas the method showing a table of the different sound fundamental frequencies can be called for a `SoundPack` instance created using an arbitrary number of `Sounds`.
 
-###### TODOOOO : rewrite example using the log-time envelop
-
-As an example, the log-time envelope, which describes the amplitude of a sound with a higher refinement in time at the start of the sound, can be computed for a sound filtered within a specific frequency bandwidth. 
+As an example, the log-time envelope, a feature describing the amplitude of a sound with a higher refinement in time at the start of the sound, can be computed for a sound filtered within a specific frequency bandwidth. 
 By plotting this feature for the same note played on two instruments, the dynamic response of the instruments can be compared for a specific frequency range.
-
-An example code snippet, which compares the Fourier transform peaks of two signals is presented below with the associated output in \autoref{fig:fft-comp}. Both sounds are loaded from wave files, conditioned by setting the signal onset at a specific value from the signal start and trimming the signal according to its fundamental. The SoundPack object was created to easily compare a set of sounds. Different features are available when instantiating the SoundPack with two and more than two sounds. The `matplotlib` Python package is used to visualize the sound features, thus the figures created by guitarsounds can be accessed to modify or save them. 
+A code snippet comparing the log-time envelope of two sounds is presented below with the associated output in \autoref{fig:log-comp}.
+In the following code, the `SoundPack` object is first instantiated from the specified sound files. 
+For each file, a `Sound` class instance is created and conditioned. 
[…]
