+++ paper/paper.bib
@@ -70,3 +70,21 @@ Subject\_term\_id: computational-neuroscience;computational-science;computer-sci
 	pages = {357--362},
 	file = {Full Text PDF:/Users/olivier/Zotero/storage/LLGGDX9M/Harris et al. - 2020 - Array programming with NumPy.pdf:application/pdf},
 }
+
+@article{hunter_matplotlib_2007,
+	title = {Matplotlib: {A} {2D} {Graphics} {Environment}},
+	volume = {9},
+	issn = {1558-366X},
+	shorttitle = {Matplotlib},
+	doi = {10.1109/MCSE.2007.55},
+	abstract = {Matplotlib is a 2D graphics package used for Python for application development, interactive scripting,and publication-quality image generation across user interfaces and operating systems},
+	number = {3},
+	journal = {Computing in Science Engineering},
+	author = {Hunter, John D.},
+	month = may,
+	year = {2007},
+	note = {Conference Name: Computing in Science Engineering},
+	keywords = {Python, application development, Computer languages, Equations, Graphical user interfaces, Graphics, Image generation, Interpolation, Operating systems, Packaging, Programming profession, scientific programming, scripting languages, User interfaces},
+	pages = {90--95},
+	file = {IEEE Xplore Abstract Record:/Users/olivier/Zotero/storage/9EB2FJTS/4160265.html:text/html;IEEE Xplore Full Text PDF:/Users/olivier/Zotero/storage/94SHQSHJ/Hunter - 2007 - Matplotlib A 2D Graphics Environment.pdf:application/pdf},
+}
