### 2023-01-17 c1f1130
debug citations #44
