+++ paper/paper.bib
@@ -21,7 +21,6 @@
 	pages = {381--383},
 }
 
-
 @article{virtanen_scipy_2020,
 	title = {{SciPy} 1.0: fundamental algorithms for scientific computing in {Python}},
 	volume = {17},
@@ -58,17 +57,8 @@
 	author = {Harris, Charles R. and Millman, K. Jarrod and van der Walt, Stéfan J. and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J. and Kern, Robert and Picus, Matti and Hoyer, Stephan and van Kerkwijk, Marten H. and Brett, Matthew and Haldane, Allan and del Río, Jaime Fernández and Wiebe, Mark and Peterson, Pearu and Gérard-Marchant, Pierre and Sheppard, Kevin and Reddy, Tyler and Weckesser, Warren and Abbasi, Hameer and Gohlke, Christoph and Oliphant, Travis E.},
 	month = sep,
 	year = {2020},
-	note = {Bandiera\_abtest: a
-Cc\_license\_type: cc\_by
-Cg\_type: Nature Research Journals
-Number: 7825
-Primary\_atype: Reviews
-Publisher: Nature Publishing Group
-Subject\_term: Computational neuroscience;Computational science;Computer science;Software;Solar physics
-Subject\_term\_id: computational-neuroscience;computational-science;computer-science;software;solar-physics},
 	keywords = {Computational neuroscience, Computational science, Computer science, Software, Solar physics},
 	pages = {357--362},
-	file = {Full Text PDF:/Users/olivier/Zotero/storage/LLGGDX9M/Harris et al. - 2020 - Array programming with NumPy.pdf:application/pdf},
 }
 
 @article{hunter_matplotlib_2007,
@@ -86,5 +76,5 @@ Subject\_term\_id: computational-neuroscience;computational-science;computer-sci
 	note = {Conference Name: Computing in Science Engineering},
 	keywords = {Python, application development, Computer languages, Equations, Graphical user interfaces, Graphics, Image generation, Interpolation, Operating systems, Packaging, Programming profession, scientific programming, scripting languages, User interfaces},
 	pages = {90--95},
-	file = {IEEE Xplore Abstract Record:/Users/olivier/Zotero/storage/9EB2FJTS/4160265.html:text/html;IEEE Xplore Full Text PDF:/Users/olivier/Zotero/storage/94SHQSHJ/Hunter - 2007 - Matplotlib A 2D Graphics Environment.pdf:application/pdf},
 }
+
