+++ paper/paper.md
@@ -1,4 +1,4 @@
---
+---
 title: 'Guitarsounds : A Python package to visualize harmonic sounds for musical instrument testing'
 tags:
   - Python
