+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Guitarsounds : A Python package to visualize harmonic sounds for musical instrument testing'
+title: 'guitarsounds : A Python package to visualize harmonic sounds for musical instrument design'
 tags:
   - Python
   - signal processing
@@ -22,16 +22,15 @@ bibliography: paper.bib
 
 # Summary
 
-The `guitarsounds` Python package can be used to rapidly visualize relevant features of harmonic sounds and is mainly developed to compare musical instrument design iterations. 
-`guitarsounds` is wrapped around the implementation of common signal processing features. 
-The implementations used in the `guitarsounds` package are manly taken from `numpy` [@harris_array_2020] and `scipy` [@virtanen_scipy_2020]. 
-The signal processing features are augmented to perform the analysis of transient harmonic sounds.
+The `guitarsounds` Python package can be used to efficiently visualize relevant features of harmonic sounds and is mainly developed to compare musical instrument design iterations. 
+`guitarsounds` is wrapped around the implementation of common signal processing features, which are manly taken from `numpy` [@harris_array_2020] and `scipy` [@virtanen_scipy_2020]. 
+The signal processing features are augmented to perform the comparative analysis of transient harmonic sounds.
 Such sounds are defined as having a defined onset, and a frequency-amplitude distribution clearly concentrated around defined partials. 
 Consequently, while the package is named `guitarsounds`, its analysis framework can be used with any harmonic sound, such as those produced by a piano, or a percussion instrument. 
 
-The guitarsounds package is divided in two main components. 
+The `guitarsounds` package is divided in two main components. 
 First, the package is developped around a convenient object oriented Application Programming Interface (API) which can be used to extract features from sounds and visualize them according to the user's needs. 
-Then, a graphical user interface (GUI) makes all the features of `guitarsounds` available to users less knowledgeable in programming. 
+Then, a graphical user interface (GUI) makes most of the features of `guitarsounds` available to users less knowledgeable in programming. 
 `guitarsounds` is meant to be used with the Jupyter Notebook interface as to allow interactively exploring the sound data, either with the API or the GUI.
 
 The main features of guitarsounds are: 
@@ -44,32 +43,32 @@ The main features of guitarsounds are:
 - Divide sounds in frequency bands to analyze variations in temporal behaviour for different frequency ranges.
 - Extract the Fourier transform peaks of a harmonic signal using a custom peak finding algorithm.
 - Extract numerical values for certain features such as the Helmholtz cavity frequency of a guitar.
-- Provide an easy-to-use signal processing API to compute new features meeting to specific needs by providing acess to lower level features and handling the differences between sound files, such as the file sample rate.
+- Provide an easy-to-use signal processing API to compute new features meeting specific needs by providing acess to lower level features and handling the differences between sound files, such as the file sample rate.
 
-Specifically, the API provides 5 classes nested together : `SoundPack`, `Sound`, `Signal`, `Plot`. 
-Starting from the innermost class, the `Plot` class handles the low level plotting of specific features, such as plotting the FFT for a single sound file.
+Specifically, the API provides 5 classes nested together : `SoundPack`, `Sound`, `Signal` and `Plot`. 
+Starting from the innermost class, the `Plot` class handles the low level plotting of specific features, such as plotting the FFT of a sound file.
 The `Signal` class is used to store the data of an array corresponding to a single signal. 
 For example, if a sound file is read and filtered, the array resulting from the filtering operation will be stored in a new instance of the `Signal` class. 
 An instance of the `Plot` class is constructed for each `Signal` class instance and stored as an attribute of the `Signal` class. 
-The `Signal` class provides all the features relying only on the data of a single sound signal as class methods. 
[…]
