### 2023-01-17 olivecha/guitarsounds#44 @olivecha
So, I've made all the changes suggested to the paper
A compiled version is available at `paper/paper.pdf`
Thanks again
