### 2023-01-17 7a11187
Revisions for SoundPack class #40
