### 2023-01-18 8610e8e
fix issues with Plot class #39
