### 2023-01-18 5c13b22
fix for Plot figures #39
