+++ guitarsounds/analysis.py
@@ -424,6 +424,8 @@ class SoundPack(object):
                 integral.append(trapezoid(log_envelop, log_time))
 
             # a list of dict for every sound
+            integral = np.array(integral)
+            integral /= np.max(integral)
             integrals.append(integral)
 
         # create the bar plotting vectors
@@ -446,6 +448,8 @@ class SoundPack(object):
                 ax.bar(x, y, width=width, tick_label=list(bin_strings), label=sound.name, color=color)
             else:
                 ax.bar(x, y, width=width, label=sound.name, color=color)
+        ax.set_xlabel('frequency bin name')
+        ax.set_ylabel('normalized power')
         plt.legend()
     
     def listen(self):
@@ -529,8 +533,8 @@ class SoundPack(object):
             plt.grid('on')
             plt.legend()
             ax = plt.gca()
-            ax.set_xlabel('Frequency (Hz)')
-            ax.set_ylabel('Amplitude (0-1)')
+            ax.set_xlabel('frequency (Hz)')
+            ax.set_ylabel('mirror amplitude (0-1)')
         else:
             print('Unsupported for multiple sounds SoundPacks')
 
@@ -552,8 +556,8 @@ class SoundPack(object):
             plt.grid('on')
             plt.plot(son1.signal.fft_frequencies()[:index], son1.signal.fft()[:index], label=son1.name)
             plt.plot(son2.signal.fft_frequencies()[:index], -son2.signal.fft()[:index], label=son2.name)
-            plt.xlabel('Fréquence (Hz)')
-            plt.ylabel('Amplitude')
+            plt.xlabel('frequency (Hz)')
+            plt.ylabel('mirror amplitude (normalized)')
             plt.title('Mirror Fourier Transform for ' + son1.name + ' and ' + son2.name)
             plt.legend()
 
@@ -591,8 +595,8 @@ class SoundPack(object):
                              label=son2.name)
             ax1.set_title('FT Histogram for ' + son1.name + ' and ' + son2.name)
             ax1.set_xscale('log')
-            ax1.set_xlabel('Fréquence (Hz)')
-            ax1.set_ylabel('Amplitude')
+            ax1.set_xlabel('frequency (Hz)')
+            ax1.set_ylabel('amplitude')
             ax1.grid('on')
             ax1.legend()
 
@@ -606,8 +610,8 @@ class SoundPack(object):
             ax2.bar(x_values[p_index], diff[p_index], width=bar_widths[p_index], color='blue', alpha=0.6)
             ax2.set_title('Difference ' + son1.name + ' - ' + son2.name)
             ax2.set_xscale('log')
-            ax2.set_xlabel('Fréquence (Hz)')
-            ax2.set_ylabel('<- Son 2 : Son 1 ->')
+            ax2.set_xlabel('frequency (Hz)')
+            ax2.set_ylabel('<- sound 2 : sound 1 ->')
             ax2.grid('on')
 
             if ticks == 'bins':
@@ -666,13 +670,17 @@ class SoundPack(object):
                 time2 = time2[:common_len]
                 integral1 = integral1[:common_len]
                 integral2 = integral2[:common_len]
+                # Normalize
+                max_value = np.max(np.hstack([integral1, integral2]))
+                integral1 /= max_value
+                integral2 /= max_value
 
                 # plot the integral area curves
                 ax.fill_between(time1, integral1, label=self.sounds[0].name, alpha=0.4)
                 ax.fill_between(time2, -integral2, label=self.sounds[1].name, alpha=0.4)
                 ax.fill_between(time2, integral1 - integral2, color='g', label='int diff', alpha=0.6)
[…]
