### 2023-01-26 6067ddd
Fix signal test
