### 2023-01-26 f6a6568
Specify ipywidgets version #33
