### 2023-01-26 olivecha/guitarsounds#33 @olivecha
It seems the 8+ version of `ipywidgets` stores the values associated to a file in a `list` of `dicts` and the version 7. in a `dict`.

`ipywidgets 7.6.5` :
```python
class FileUpload(DescriptionWidget, ValueWidget, CoreWidget):
    """
    Upload file(s) from browser to Python kernel as bytes
    """
    [...]
    error = Unicode(help='Error message').tag(sync=True)
    value = Dict(read_only=True)
```
`ipywidgets 8.0.4`
```python
class FileUpload(DescriptionWidget, ValueWidget, CoreWidget):
    """File upload widget

    This creates a file upload input that allows the user to select
    one or more files to upload. The file metadata and content
    can be retrieved in the kernel.

    Examples
    --------

    >>> import ipywidgets as widgets
    >>> uploader = widgets.FileUpload()

    # After displaying `uploader` and uploading a file:

    >>> uploader.value
    [
      {
        'name': 'example.txt',
        'type': 'text/plain',
        'size': 36,
        'last_modified': datetime.datetime(2020, 1, 9, 15, 58, 43, 321000, tzinfo=datetime.timezone.utc),
        'content': <memory at 0x10c1b37c8>
      }
    ]
    >>> uploader.value[0].content.tobytes()
    b'This is the content of example.txt.\n'
```

I've specified the `ipywidgets` version in `requirements.txt`
