### 2023-01-30 833c697
Corrections to Signal class #38
