+++ guitarsounds/analysis.py
@@ -784,10 +784,8 @@ class Sound(object):
         """
         A method conditioning the Sound instance.
         - Trimming to just before the onset
-        - Filtering the noise
         :param verbose: if True problem with trimming and filtering are reported
         :param return_self: If True the method returns the conditioned Sound instance
-        :param filter_noise: If True the Sound is filtered using a noise reducing algorithm
         :param auto_trim: If True, the sound is trimmed to a fixed length according to its fundamental
         :return: a conditioned Sound instance if `return_self = True`
         """
@@ -836,15 +834,6 @@ class Sound(object):
         self.bins = self.signal.make_freq_bins()
         # unpack the bins
         self.bass, self.mid, self.highmid, self.uppermid, self.presence, self.brillance = self.bins.values()
-
-    def filter_noise(self, verbose=True):
-        """
-        Filters the noise in the signal attribute
-        :param verbose: if True problem are printed to the terminal
-        :return: None
-        """
-        # filter the noise in the Signal class
-        self.signal = self.trimmed_signal.filter_noise(verbose=verbose)
 
     def trim_signal(self, verbose=True):
         """
@@ -953,17 +942,18 @@ class Signal(object):
         :param signal: vector containing the signal samples
         :param sr: sample rate of the signal (Hz)
         :param SoundParam: Sound Parameter instance to use with the signal
-        :para freq_range: Frequency range to use with the signal
+        :para freq_range: Frequency range to use with the signal.
+        This is the maximum frequency used when computing Fourier transform
+        peaks and plotting Fourier transform related plots.
         """
         self.SP = SoundParam
         self.onset = None
         self.signal = signal
         self.sr = sr
         self.range = freq_range
-        self.trimmed = None
-        self.noise = None
         self.plot = Plot()
         self.plot.parent = self
+        self.norm_factor = None
 
     def time(self):
         """
@@ -1093,8 +1083,7 @@ class Signal(object):
         :return: The damping ratio, a scalar.
         """
         # Get the envelop data
-        envelop_time = self.normalize().envelop_time()
-        envelop = self.normalize().envelop()
+        envelop, envelop_time = self.parent.normalize().envelop() 
 
         # First point is the maximum because e^-kt is strictly decreasing
         first_index = np.argmax(envelop)
@@ -1163,7 +1152,7 @@ class Signal(object):
         Finds the Hemlotz cavity frequency index from the Fourier Transform by searching for a peak in the expected
         range (80 - 100 Hz), if the fundamental is too close to the expected Hemlotz frequency a comment
         is printed and None is returned.
-        :return: If successful the cavity peak index
+        :return: The index of the cavity peak
         """
         first_index = np.where(self.fft_frequencies() >= 80)[0][0]
         second_index = np.where(self.fft_frequencies() >= 110)[0][0]
@@ -1195,9 +1184,13 @@ class Signal(object):
 
     def fft_bins(self):
         """
-        Transforms the Fourier Transform signal into a statistic distribution.
-        Occurrences of each frequency varies from 0 to 100 according to their
-        amplitude.
+        Transforms the Fourier transform array into a statistic distribution 
+        ranging from 0 to 100. 
+        Accordingly, the maximum of the fourier transform with value 1.0 will
+        be equal to 100 and casted as an integer.
+        Values below 0.001 will be equal to 0. 
+        This representation of the Fourier transform is used to construct
+        octave bands histograms.
         :return : a list containing the frequency occurrences.
         """
 
@@ -1215,7 +1208,15 @@ class Signal(object):
     def envelop(self, window=None, overlap=None):
         """
[…]
