### 2023-01-30 cb27dc0
fix typos in README.md #32
