### 2023-01-30 326152b
Refractor Sound class #37
