### 2023-01-30 284abe3
Refractor Sound class #37
