### 2023-01-30 76828ef
Remove Sound.condition statements #37
