### 2023-02-01 7eb5fec
Move sound_parameter call inside class
