+++ guitarsounds/__init__.py
@@ -1,8 +1,6 @@
 # __init__.py
 from guitarsounds.interface import guitarGUI as Interface
 from guitarsounds.analysis import Signal, Sound, SoundPack, Plot
-from guitarsounds.analysis import plt
-show = plt.show
 
 # Version of the guitarsounds package
-__version__ = '1.0.0'
+__version__ = '1.1.0'
