+++ guitarsounds/utils.py
@@ -10,6 +10,7 @@ from guitarsounds.parameters import sound_parameters
 
 # instantiation of the trim time interpolator
 sp = sound_parameters()
+
 freq_dict = {'E2': 82.41,
              'A2': 110.0,
              'D3': 146.83,
@@ -178,3 +179,12 @@ def resample(y, sr_orig, sr_target=22050):
     y_len = int(sr_target * len(y) / sr_orig)
     y_new = scipy.signal.resample(y, num=y_len)
     return y_new
+
+
+def generate_error_widget(text):
+    """
+    Creates a ipywidget error message from a string
+    :param text: python str of the error message
+    :return: widgets.HTLM object corresponding to the error message
+    """
+    return widgets.HTML('<p style="color:#CC4123;">' + text + '</p>')
