### 2023-02-01 28640e1
Update static.yml
