+++ .github/workflows/static.yml
@@ -4,7 +4,7 @@ name: Deploy static content to Pages
 on:
   # Runs on pushes targeting the default branch
   push:
-    branches: ["main"]
+    branches: ["main", "JOSS-Paper"]
 
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
