### 2023-02-01 4d21ecd
Update static.yml
