+++ guitarsounds/analysis.py
@@ -5,28 +5,21 @@ import matplotlib.pyplot as plt
 import matplotlib.cm
 import numpy as np
 import os
-# from noisereduce.noisereducev1 import reduce_noise
 import scipy
 import scipy.optimize
 import scipy.integrate
 import scipy.interpolate
+from scipy.integrate import trapezoid
 from scipy import signal as sig
 from guitarsounds.parameters import sound_parameters
 import guitarsounds.utils as utils
 from tabulate import tabulate
 import wave
 
-wave_ver = wave
-
-"""
-Getting the sound parameters from the guitarsounds_parameters.py file
-"""
-SP = sound_parameters()
 
 """
 Classes
 """
-
 
 class SoundPack(object):
     """
@@ -61,7 +54,7 @@ class SoundPack(object):
         """
         # create a copy of the sound parameters
         if SoundParams is None:
-            self.SP = SP
+            self.SP = sound_parameters()
         else:
             self.SP = SoundParams
 
@@ -78,9 +71,11 @@ class SoundPack(object):
             # general case for multiple sounds
             self.kind = 'multiple'
 
+        # If filenames are supplied
         if type(sounds[0]) is str:
             self.sounds_from_files(sounds, names=names, fundamentals=fundamentals)
 
+        # Else sound instances are supplied
         else:
             self.sounds = sounds
 
@@ -103,10 +98,12 @@ class SoundPack(object):
                     for sound, n in zip(self.sounds, names):
                         sound.name = n
 
-        if equalize_time:
+            # If the sounds are not conditionned condition them
             for s in self.sounds:
                 if ~hasattr(s, 'signal'):
                     s.condition()
+
+        if equalize_time:
             self.equalize_time()
 
         # Define bin strings
@@ -136,7 +133,7 @@ class SoundPack(object):
         self.sounds = []
         for file, name, fundamental in zip(sound_files, names, fundamentals):
             self.sounds.append(Sound(file, name=name, fundamental=fundamental,
-                                     SoundParams=self.SP).condition(return_self=True))
+                                     SoundParams=self.SP))
 
     def equalize_time(self):
         """
@@ -165,7 +162,9 @@ class SoundPack(object):
             sound.signal = sound.signal.normalize()
             new_sounds.append(sound)
 
-        return SoundPack(new_sounds, names=names, fundamentals=fundamentals, SoundParams=self.SP, equalize_time=False)
+        self.sounds = new_sounds
+
+        return self
 
     """
     Methods for all SoundPacks
@@ -288,7 +287,7 @@ class SoundPack(object):
                 plt.title(title0 + title1 + title2)
             plt.tight_layout()
 
-        elif f_bin in [*list(SP.bins.__dict__.keys())[1:], 'brillance']:
+        elif f_bin in [*list(sound_parameters().bins.__dict__.keys())[1:], 'brillance']:
             plt.figure(figsize=(10, 4))
             # Plot every envelop for a single frequency bin
             norm_factors = np.array([son.bins[f_bin].normalize().norm_factor for son in self.sounds])
@@ -305,90 +304,6 @@ class SoundPack(object):
 
         else:
             print('invalid frequency bin')
-
-    def combine_envelop(self, kind='signal', difference_factor=1, show_sounds=True, show_rejects=True, **kwargs):
-        """
-        __ Multiple SoundPack Method __
-        Combines the envelopes of the Sounds contained in the SoundPack, Sounds having a too large difference factor
-        from the average are rejected.
-        :param kind: which signal to use from :
-        'signal', 'bass', 'mid', 'highmid', 'uppermid', 'presence', 'brillance'
[…]
