+++ tests/test_SoundPack.py
@@ -1,6 +1,9 @@
 import unittest
+from contextlib import redirect_stdout
+from io import StringIO
 import guitarsounds
-from guitarsounds import SoundPack, helpers_tests
+from guitarsounds import SoundPack
+import helpers_tests
 import numpy as np
 import matplotlib.axes
 import matplotlib.pyplot as plt
@@ -43,7 +46,7 @@ class MyTestCase(unittest.TestCase):
         sp = SoundPack(files)
         nsp = sp.normalize()
         for s in nsp.sounds:
-            self.assertTrue(abs(np.max(np.abs(s.signal.signal)) - 1.) < 1e-3)
+            self.assertAlmostEqual(np.max(np.abs(s.signal.signal)) , 1., 3)
 
     def test_SoundPack_plot(self):
         files = [helpers_tests.get_rnd_audio_file() for _ in range(3)]
@@ -81,7 +84,124 @@ class MyTestCase(unittest.TestCase):
             plt.close(plt.gcf())            
     
     def test_SoundPack_listen(self):
+        """ Test the SoundPack.listen method """
         files = [helpers_tests.get_rnd_audio_file() for _ in range(4)]
         sp = SoundPack(files)
         sp.listen()
         self.assertTrue(True)
+
+    def test_soundpack_freq_bin_plot(self):
+        """ test the SoundPack.freq_bin_plot method """
+        # Create a SoundPack instance
+        files = [helpers_tests.get_rnd_audio_file() for _ in range(3)]
+        sp = SoundPack(files)
+        # Test branch with all frequency bins
+        sp.freq_bin_plot(f_bin='all')
+        # test that all the frequency bins are plotted
+        self.assertTrue(len(plt.gcf().axes) == 6)
+        plt.close(plt.gcf())
+        # Test branch with a specific frequency bin
+        sp.freq_bin_plot(f_bin='mid')
+        self.assertTrue(len(plt.gcf().axes) == 1)
+
+    def test_soundpack_fundamentals(self):
+        """ test the fundamentals method of SoundPack """
+        # create a SoundPack instance
+        files = [helpers_tests.get_rnd_audio_file() for i in range(3)]
+        sp = SoundPack(files)
+        # redirect_stdout
+        output = StringIO()
+        with redirect_stdout(output):
+            sp.fundamentals()
+        self.assertTrue(len(output.getvalue()) > 0)
+
+    def test_soundpack_integral_plot(self):
+        """ test the SoundPack integral_plot method """
+        # Create a soundpack
+        files = [helpers_tests.get_rnd_audio_file() for i in range(3)]
+        sp = SoundPack(files)
+        # Test branch with all frequency bins
+        sp.integral_plot(f_bin='all')
+        self.assertTrue(len(plt.gcf().axes) == 6)
+        # Test branch with a specific frequency bin
+        sp.integral_plot(f_bin='highmid')
+        self.assertTrue(len(plt.gcf().axes) == 1)
+
+    def test_soundpack_bin_power_table(self):
+        """ Test the SoundPack bin_power_table method """
+        # create a SoundPack instance
+        files = [helpers_tests.get_rnd_audio_file() for i in range(3)]
+        sp = SoundPack(files)
+        # redirect_stdout
+        output = StringIO()
+        with redirect_stdout(output):
+            sp.bin_power_table()
+        self.assertTrue(len(output.getvalue()) > 0)
+
+    def test_soundpack_bin_power_hist(self):
+        """ Test the bin_power_hist method of SoundPack """
+        # create a SoundPack instance
+        files = [helpers_tests.get_rnd_audio_file() for i in range(3)]
+        sp = SoundPack(files)
+        # plot the bin power histogram
+        sp.bin_power_hist()
[…]
