+++ .github/workflows/python-app.yml
@@ -37,5 +37,4 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
       run: |
-        cd tests
-        pytest
+        pytest tests
