+++ README.md
@@ -3,15 +3,15 @@
 [![pytest](https://github.com/olivecha/guitarsounds/actions/workflows/python-app.yml/badge.svg)](https://github.com/olivecha/guitarsounds/actions/workflows/python-app.yml) 
 
 A python package to analyse guitar sounds. Developed as a lutherie research analysis tool with the [Bruand Lutherie School](https://bruand.com/). 
-The guitarsound python package documentation is available at [documentation](https://olivecha.github.io/guitarsounds/).
+The guitarsound python package documentation is available at [documentation](https://olivecha.github.io/guitarsounds/) and a Jupyter Notebook tutorial for the API is available [here](https://github.com/olivecha/guitarsounds/blob/JOSS-Paper/docs/API_Tutorial.ipynb).
 
 ## Motivation
 
-The main goal of this project is to provide a tool to efficiently analyse sound data from research projects in musical instrument desing. While sound analysis packages already exist, they are more directed to feature extraction for machine learning purposes. Additionnaly, some features of interest, like time dependent decay, onset shape and fourier transform peaks distribution are not computable trivially or acurately with existing tools. The current release of the guitarsounds package contains usual and advanced digital signal processing tools applied to the analysis of transient harmonic sounds with easy figure generation trough `matplotlib`. To allow the package functionalities to be used rapidly without learning the API, a graphic user interface is available based on jupyter lab widgets.
+The main goal of this project is to provide a tool to efficiently analyse sound data from research projects in musical instrument design. While sound analysis packages already exist, they are more directed to feature extraction for machine learning purposes. Additionally, some features of interest, like time dependent decay, onset shape and fourier transform peaks distribution are not computable trivially or accurately with existing tools. The current release of the guitarsounds package contains usual and advanced digital signal processing tools applied to the analysis of transient harmonic sounds with easy figure generation through `matplotlib`. To allow the package functionalities to be used rapidly without learning the API, a graphic user interface is available based on jupyter lab widgets.
 
 ## Installation
 
-The following steps can be followed to use guitarsound interactively with Jupyter Notebook. A french version of the installation guide is available [here](https://github.com/olivecha/guitarsounds/wiki/Guide-installation-fran%C3%A7ais).
+The following steps can be followed to use guitarsound interactively with Jupyter Notebook. A French version of the installation guide is available [here](https://github.com/olivecha/guitarsounds/wiki/Guide-installation-fran%C3%A7ais).
 
 - Download the Anaconda package management system [link](https://www.anaconda.com/products/distribution).
 
@@ -19,7 +19,7 @@ The following steps can be followed to use guitarsound interactively with Jupyte
 
 - Install the Anaconda package management system ([Tutorial](https://docs.anaconda.com/anaconda/install/windows/)).
 
-- Once Anaconda is installed, the guitarsound package needs to be installed. Without going into heavy details, the guitarsound package is not available from the Anaconda channels and needs to be installed using PIP, the [Package Installer for Python](https://pypi.org/project/guitarsounds/). In order to install a package from PIP in Anaconda, you need to use the Anaconda [Prompt](https://docs.anaconda.com/anaconda/user-guide/getting-started/#open-prompt-win) :
[…]
