### 2023-02-01 olivecha/guitarsounds#36 @olivecha
- I've added a github action to solve issue #29
- the `:param :` syntax appears to compile correctly on the docs [webpage](https://olivecha.github.io/guitarsounds/guitarsounds/analysis.html)
- Corrected typos in code and documentation
