### 2023-02-01 #4878 @olivecha
@danielskatz all the issues raised by the reviewers have been addressed! Only minor work is needed for the issue [36](https://github.com/olivecha/guitarsounds/issues/36) to improve the documentation with explanations about the available features.
