+++ guitarsounds/analysis.py
@@ -115,7 +115,9 @@ class SoundPack(object):
 
     def sounds_from_files(self, sound_files, names=None, fundamentals=None):
         """
-        Create Sound class instances and assign them to the SoundPack from a list of files
+        Create Sound class instances and assign them to the SoundPack 
+        from a list of files
+
         :param sound_files: sound filenames
         :param names: sound names
         :param fundamentals: user specified fundamental frequencies
@@ -137,7 +139,8 @@ class SoundPack(object):
 
     def equalize_time(self):
         """
-        Trim the sounds so that they all have the length of the shortest sound, trimming is done at the end.
+        Trim the sounds so that they all have the length of the shortest sound, 
+        trimming is done at the end of the sounds.
         :return: None
         """
         trim_index = np.min([len(sound.signal.signal) for sound in self.sounds])
@@ -152,7 +155,7 @@ class SoundPack(object):
     def normalize(self):
         """
         Normalize all the signals in the SoundPack and returns a normalized
-        instance of itself
+        instance of itself. See `Signal.normalize` for more information.
         :return: SoundPack with normalized signals
         """
         new_sounds = []
@@ -176,10 +179,13 @@ class SoundPack(object):
 
         __ Multiple SoundPack Method __
         Plots a specific signal.plot for all sounds on the same figure
-        Ex : compare_plot('fft') plots the fft of all sounds on a single figure
+
+        Ex : SoundPack.plot('fft') plots the fft of all sounds on a single figure
         The color argument is set to none so that the plots have different colors
-        :param kind: Attribute passed to the `signal.plot()` method
-        :param kwargs: keywords arguments to pass to the `signal.plot()` method
+        :param kind: feature name passed to the `signal.plot()` method
+        :param kwargs: keywords arguments to pass to the `matplotlib.plot()` 
+        method
+
         :return: None
         """
         plt.figure(figsize=(8, 6))
@@ -198,11 +204,12 @@ class SoundPack(object):
 
         __ Multiple SoundPack Method __
         Draws the same kind of plot on a different axis for each sound
-        Example : `SoundPack.compare_plot('peaks')` with 4 Sounds will plot a figure with 4 axes, with each
-        a different 'peak' plot.
+        Example : `SoundPack.compare_plot('peaks')` with 4 Sounds will plot a 
+        figure with 4 axes, with each a different 'peak' plot.
 
         :param kind: kind argument passed to `Signal.plot()`
-        :param kwargs: key word arguments passed to Signal.plot()
+        :param kwargs: key word arguments passed to `matplotlib.plot()`
+
         :return: None
         """
         # if a dual SoundPack : only plot two big plots
@@ -254,14 +261,15 @@ class SoundPack(object):
 
     def freq_bin_plot(self, f_bin='all'):
         """
-        Plots the log envelop of specified frequency bins
+        Plots the log envelope of specified frequency bins
 
         __ Multiple SoundPack Method __
-        A function to compare signals decomposed frequency wise in the time domain on a logarithm scale.
-        The methods plot all the sounds and plots their frequency bins according to the frequency bin argument f_bin.
+        A function to compare signals decomposed frequency wise in the time 
+        domain on a logarithm scale. The methods plot all the sounds and plots 
+        their frequency bins according to the frequency bin argument f_bin.
 
-        Example : SoundPack.freq_bin_plot(f_bin='mid') will plot the log-scale envelop of the 'mid' signal of every
-        sound in the SoundPack
+        Example : SoundPack.freq_bin_plot(f_bin='mid') will plot the log-scale 
+        envelope of the 'mid' signal of every sound in the SoundPack.
 
[…]
