### 2023-02-01 dda90b1
Update static.yml
