+++ .github/workflows/static.yml
@@ -35,6 +35,7 @@ jobs:
         sudo apt-get install libsndfile1-dev
         python -m pip install --upgrade pip
         pip install pdoc
+        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Update docs
       run: |
         pdoc -o docs guitarsounds
