+++ tests/test_Signal.py
@@ -87,7 +87,7 @@ class MyTestCase(unittest.TestCase):
         rnd_sig = get_rnd_test_Signal()
         # higher frequency is less damped
         freq_damping = rnd_sig.peak_damping()
-        self.assertTrue(freq_damping[-1] < freq_damping[0])
+        self.assertTrue(freq_damping[-1] < freq_damping[1])
         # test for known signal
         ref_sig = get_ref_test_Signal()
         freq_damping = ref_sig.peak_damping()
