### 2023-02-01 a8823dc
Update static.yml
