+++ .github/workflows/static.yml
@@ -37,6 +37,8 @@ jobs:
         pip install pdoc
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Update docs
+      env:
+        PDOC_ALLOW_EXEC: 1
       run: |
         pdoc -o docs guitarsounds
   #  job to deploy to github pages
