### 2023-02-01 032f7aa
Update static.yml
