### 2023-02-01 f4f7604
Update static.yml
