### 2023-02-01 a6f13db
Update static.yml
