+++ .github/workflows/static.yml
@@ -38,7 +38,7 @@ jobs:
         pdoc -o docs guitarsounds
         
     - name: Commit and push changes
-    - uses: stefanzweifel/git-auto-commit-action@v4
+      uses: stefanzweifel/git-auto-commit-action@v4
       with:
         commit_message: Update docs
         
