### 2023-02-01 263238e
Update static.yml
