+++ .github/workflows/static.yml
@@ -11,9 +11,12 @@ on:
 
 # Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
 permissions:
-  contents: read
+  contents: write
   pages: write
   id-token: write
+  pull-requests: write
+  issues: write
+  repository-projects: write
 
 jobs:
   #  job to update the docs
@@ -48,6 +51,7 @@ jobs:
       name: github-pages
       url: ${{ steps.deployment.outputs.page_url }}
     runs-on: ubuntu-latest
+    needs: update
     steps:
       - name: Checkout
         uses: actions/checkout@v3
