### 2023-02-01 ff08443
Update static.yml
