+++ guitarsounds/analysis.py
@@ -24,28 +24,45 @@ Classes
 class SoundPack(object):
     """
     A class to store and analyse multiple sounds
-    Some methods are only available for the case with two sounds
+    Some methods are only available for SoundPacks containing two sounds
     """
 
-    def __init__(self, *sounds, names=None, fundamentals=None, SoundParams=None, equalize_time=True):
+    def __init__(self, *sounds, names=None, fundamentals=None, 
+                 SoundParams=None, equalize_time=True):
         """
-        The SoundPack can be instantiated from existing Sound class instances, either in a list or as
-        multiple arguments
+        The SoundPack can be instantiated from existing Sound class instances, 
+        either in a list or as multiple arguments
+        The class can also handle the creation of Sound class instances if the 
+        arguments are filenames, either a list or multiple arguments.
 
-        The class can also handle the creation of Sound class instances if the arguments are filenames,
-        either a list or multiple arguments.
+        :param sounds: `guitarsounds.Sound` instaces or filenames either as 
+        multiple arguments or as a list
+        :param names: list of strings with the names of the sounds that will be 
+        used in the plot legend labels 
+        :param fundamentals: list of numbers corresponding to the known sound 
+        fundamental frequencies. 
+        :param SoundParams: `guitarsounds.SoundParams` instance used to get 
+        the parameters used in the computation of the sound attributes 
+        :param equalize_time: if True, all the sounds used to create the 
+        SoundPack are truncated to the length of the shortest sound. 
 
-        If the number of Sound contained is equal to two, the SoundPack will be 'dual' and the associated methods
-        will be available
+        If the number of Sound contained is equal to two, the SoundPack will 
+        be 'dual' and the associated methods will be available :
 
-        If it contains multiple sounds the SoundPack will be multiple and a reduced number of methods will work
+            - `SoundPack.compare_peaks`
+            - `SoundPack.fft_mirror`
+            - `SoundPack.fft_diff`
+            - `SoundPack.integral_compare`
 
-        A list of names as strings and fundamental frequencies can be specified when creating the SoundPack
+        If it contains multiple sounds the SoundPack will be multiple and a 
+        reduced number of methods will be available to call
 
-        If equalize_time is set to False, the contained sounds will not be trimmed to the same length.
+        If the fundamental frequency is supplied for each sound, the 
+        computation of certain features can be more efficient, such as the 
+        time damping computation or Hemlotz frequency computation.
 
         Examples :
-        ```
+        ```python
         Sound_Test = SoundPack('sounds/test1.wav', 'sounds/test2.wav', names=['A', 'B'], fundamentals = [134, 134])
 
         sounds = [sound1, sound2, sound3, sound4, sound5] # instances of the Sound class
@@ -177,16 +194,16 @@ class SoundPack(object):
         """
         Superimposed plot of all the sounds on one figure for a specific kind
 
+        :param kind: feature name passed to the `signal.plot()` method
+        :param kwargs: keywords arguments to pass to the `matplotlib.plot()` 
+        method
+        :return: None
+
         __ Multiple SoundPack Method __
         Plots a specific signal.plot for all sounds on the same figure
 
         Ex : SoundPack.plot('fft') plots the fft of all sounds on a single figure
         The color argument is set to none so that the plots have different colors
-        :param kind: feature name passed to the `signal.plot()` method
-        :param kwargs: keywords arguments to pass to the `matplotlib.plot()` 
-        method
-
-        :return: None
         """
         plt.figure(figsize=(8, 6))
[…]
