### 2023-02-04 olivecha/guitarsounds#33 @cwitkowitz
This is indeed fixed by installing `ipywidgets==7.6.5`, however I don't see the referenced updates to `requirements.txt`. I will close the issue after that update is made.

You may also consider adding `jupyter` to `requirements.txt`, as I needed to install it manually in order to run `jupyter notebook` from the command line.
