### 2023-02-04 olivecha/guitarsounds#44 @cwitkowitz
Thank you for addressing all of my comments. In my opinion, the quality of the paper has improved significantly.

Here are a few more minor comments after reading the revisions:
- In the paragraph before the code snippet, break apart the sentence "... equal to 22050 Hz, this is important ..." into two sentences (at the comma)
- At least one instance of guitarsounds remains as text and not a code snippet (just double-check these)
- "manly" -> "mainly"
- "developped" -> "developed" (appears twice)
- "acess" -> "access"
- "two guitar designed" -> "two guitar designs"
- "Hemlotz" -> "Helmholtz"
- "machine learning application" -> "machine learning applications"

If you make these minor adjustments, I would be happy to close the issue (assuming @ebezzam is also satisfied with the revisions).
