### 2023-02-04 #4878 @cwitkowitz
Thank you @olivecha. I am very satisfied with your revisions and appreciate your attention to detail in addressing all of my comments. I can confidently mark off the remaining checklist items.

Currently, only the following issues I raised remain:
- https://github.com/olivecha/guitarsounds/issues/33
  Just make sure `requirements.txt` is updated.
- https://github.com/olivecha/guitarsounds/issues/36
  It would be nice to have short descriptions of less well-known features in the documentation.
- https://github.com/olivecha/guitarsounds/issues/44
  We can close this after @ebezzam has a chance to take a look.

However, I don't feel these block acceptance at this point (besides 44 obviously).
