+++ guitarsounds/analysis.py
@@ -304,8 +304,8 @@ class SoundPack(object):
                 plt.xscale('log')
                 plt.legend()
                 son = self.sounds[-1]
-                title0 = ' ' + key + ' : ' + str(int(son.bins[key].range[0])) + ' - ' + str(
-                         int(son.bins[key].range[1])) + ' Hz, '
+                title0 = ' ' + key + ' : ' + str(int(son.bins[key].freq_range[0])) + ' - ' + str(
+                         int(son.bins[key].freq_range[1])) + ' Hz, '
                 title1 = 'Norm. Factors : '
                 title2 = 'x, '.join(str(np.around(norm_factor, 0)) for norm_factor in norm_factors)
                 plt.title(title0 + title1 + title2)
@@ -320,8 +320,8 @@ class SoundPack(object):
             plt.xscale('log')
             plt.legend()
             son = self.sounds[-1]
-            title0 = ' ' + f_bin + ' : ' + str(int(son.bins[f_bin].range[0])) + ' - ' + str(
-                int(son.bins[f_bin].range[1])) + ' Hz, '
+            title0 = ' ' + f_bin + ' : ' + str(int(son.bins[f_bin].freq_range[0])) + ' - ' + str(
+                int(son.bins[f_bin].freq_range[1])) + ' Hz, '
             title1 = 'Norm. Factors : '
             title2 = 'x, '.join(str(np.around(norm_factor, 0)) for norm_factor in norm_factors)
             plt.title(title0 + title1 + title2)
@@ -371,8 +371,8 @@ class SoundPack(object):
                     sound.bins[key].plot.integral(label=sound.name)
                 plt.legend()
                 sound = self.sounds[-1]
-                title0 = ' ' + key + ' : ' + str(int(sound.bins[key].range[0])) + ' - ' + str(
-                    int(sound.bins[key].range[1])) + ' Hz, '
+                title0 = ' ' + key + ' : ' + str(int(sound.bins[key].freq_range[0])) + ' - ' + str(
+                    int(sound.bins[key].freq_range[1])) + ' Hz, '
                 title1 = 'Norm. Factors : '
                 title2 = 'x, '.join(str(np.around(norm_factor, 0)) for norm_factor in norm_factors)
                 plt.title(title0 + title1 + title2)
@@ -387,8 +387,8 @@ class SoundPack(object):
                 sound.bins[f_bin].plot.integral(label=sound.name)
             plt.legend()
             sound = self.sounds[-1]
-            title0 = ' ' + f_bin + ' : ' + str(int(sound.bins[f_bin].range[0])) + ' - ' + str(
-                int(sound.bins[f_bin].range[1])) + ' Hz, '
+            title0 = ' ' + f_bin + ' : ' + str(int(sound.bins[f_bin].freq_range[0])) + ' - ' + str(
+                int(sound.bins[f_bin].freq_range[1])) + ' Hz, '
             title1 = 'Norm. Factors : '
             title2 = 'x, '.join(str(np.around(norm_factor, 0)) for norm_factor in norm_factors)
             plt.title(title0 + title1 + title2)
@@ -779,31 +779,33 @@ class SoundPack(object):
 
 class Sound(object):
     """
-    A class to store audio signals obtained from a sound and compare them
+    A class to store audio signals corresponding to a sound and compute 
+    features on them.
     """
 
-    def __init__(self, data, name='', condition=True, auto_trim=False,
-                 use_raw_signal=False, normalize_raw_signal=False, 
-                 fundamental=None, SoundParams=None):
+    def __init__(self, data, name='', fundamental=None, condition=True, 
+                 auto_trim=False, use_raw_signal=False, 
+                 normalize_raw_signal=False, SoundParams=None):
         """
-        Creates a Sound instance from a .wav file, name as a string and 
[…]
