### 2023-02-06 olivecha/guitarsounds#33 @olivecha
The requirements.txt were only update on the JOSS branch, I've merged this change on the main branch and added jupyter to requirements.txt.
