+++ paper/paper.md
@@ -23,17 +23,17 @@ bibliography: paper.bib
 # Summary
 
 The `guitarsounds` Python package can be used to efficiently visualize relevant features of harmonic sounds and is mainly developed to compare musical instrument design iterations. 
-`guitarsounds` is wrapped around the implementation of common signal processing features, which are manly taken from `numpy` [@harris_array_2020] and `scipy` [@virtanen_scipy_2020]. 
+`guitarsounds` is wrapped around the implementation of common signal processing features, which are mainly taken from `numpy` [@harris_array_2020] and `scipy` [@virtanen_scipy_2020]. 
 The signal processing features are augmented to perform the comparative analysis of transient harmonic sounds.
 Such sounds are defined as having a clear onset, and a frequency-amplitude distribution clearly concentrated around partials. 
 Consequently, while the package is named `guitarsounds`, its analysis framework can be used with any harmonic sound, such as those produced by a piano, or a percussion instrument. 
 
 The `guitarsounds` package is divided in two main components. 
-First, the package is developped around a convenient object-oriented Application Programming Interface (API) which can be used to extract features from sounds and visualize them according to the user's needs. 
+First, the package is developed around a convenient object-oriented Application Programming Interface (API) which can be used to extract features from sounds and visualize them according to the user's needs. 
 Then, a graphical user interface (GUI) makes most of the features of `guitarsounds` available to users less knowledgeable in programming. 
 `guitarsounds` is meant to be used with the Jupyter Notebook interface as to allow interactively exploring the sound data, either with the API or the GUI.
 
-The main features of guitarsounds are: 
+The main features of `guitarsounds` are: 
 
 - Automate the loading, conditioning and normalization of multiple sound files to meaningfully compare their features.
 - Visualize sounds features relevant to musical instrument design, such as: 
@@ -43,7 +43,7 @@ The main features of guitarsounds are:
 - Divide sounds in frequency bands to analyze variations in temporal behaviour for different frequency ranges.
 - Extract the Fourier transform peaks of a harmonic signal using a custom peak finding algorithm.
 - Extract numerical values for certain features such as the Helmholtz cavity frequency of a guitar.
-- Provide an easy-to-use signal processing API to compute new features meeting specific needs by providing acess to lower level features and handling the differences between sound files, such as the file sample rate.
+- Provide an easy-to-use signal processing API to compute new features meeting specific needs by providing access to lower level features and handling the differences between sound files, such as the file sample rate.
 
 Specifically, the API provides 5 classes nested together : `SoundPack`, `Sound`, `Signal` and `Plot`. 
 Starting from the innermost class, the `Plot` class handles the low level plotting of specific features, such as plotting the Fast Fourier Transform (FFT) of a sound file.
@@ -63,7 +63,7 @@ By plotting this feature for the same note played on two instruments, the dynami
 A code snippet comparing the log-time envelope of two sounds is presented below with the associated output in \autoref{fig:log-comp}.
 In the following code, the `SoundPack` object is first instantiated from the specified sound files. 
 For each file, a `Sound` class instance is created and conditioned. 
-In the conditioning procedure, the signal is first resampled to have a sample rate equal to 22050 Hz, this is important to ensure all the features compared between sounds are computed using the same sample rate. 
+In the conditioning procedure, the signal is first resampled to have a sample rate equal to 22050 Hz. This is important to ensure all the features compared between sounds are computed using the same sample rate. 
 The sound is then trimmed so that the beginning of the onset is at 100 ms, as can be seen on \autoref{fig:log-comp}. 
 To ensure compared sounds have the same length, the ends of the sounds are trimmed so that each sound has the same number of samples as the shortest sound. 
[…]
