### 2023-02-06 olivecha/guitarsounds#44 @olivecha
I've made the minor adjustments and added an updated version of paper.pdf to the JOSS branch for convenience.
