### 2023-02-07 olivecha/guitarsounds#44 @ebezzam
I agree the paper has significantly improved.

A few comments / suggestions:
- Line 38, you list 4 classes.
- (Suggestion) Lines 39 - 57 as a list to improve readability. E.g.
  - `SoundPack`: description...
  - `Sound`: description...
  - etc...
- Line 77 "guitar" -> "guitars"

Technical comment:
- Line 68-69: why not extend the shortest sound so that the two audio have the same length? I suppose it could be interesting to see how much longer the decay is for a specific instrument?
