### 2023-02-07 #4878 @ebezzam
Thank you @olivecha for the updates! Everything checks off for me, but with couple things to address:

- Some final comments in https://github.com/olivecha/guitarsounds/issues/44
- One of the unit tests failed for me ([link](https://github.com/olivecha/guitarsounds/blob/41215ac9c8f5b13b7c25ad5d8e0fa3171126e6fb/tests/test_Signal.py#L90) to code), but only the first time I ran it (traceback below). I guess because I came across a random initialization that broke the test. So perhaps fixing the seed might be useful?

```
tests/test_Plot.py ............                                                                                                                                                                             [ 14%]
tests/test_Signal.py ...........F..........                                                                                                                                                                 [ 41%]
tests/test_Sound.py ..........                                                                                                                                                                              [ 53%]
tests/test_SoundPack.py ................                                                                                                                                                                    [ 73%]
tests/test_interface.py ...............                                                                                                                                                                     [ 91%]
tests/test_utils.py .......                                                                                                                                                                                 [100%]

==================================================================================================== FAILURES =====================================================================================================
_______________________________________________________________________________________ MyTestCase.test_Signal_peak_damping _______________________________________________________________________________________

self = <test_Signal.MyTestCase testMethod=test_Signal_peak_damping>

    def test_Signal_peak_damping(self):
        """ Test the peak damping method of the signal"""
        # test for random signal
        rnd_sig = get_rnd_test_Signal()
        # higher frequency is less damped
        freq_damping = rnd_sig.peak_damping()
>       self.assertTrue(freq_damping[-1] < freq_damping[1])
E       AssertionError: False is not true

tests/test_Signal.py:90: AssertionError
```
