### 2023-02-07 olivecha/guitarsounds#44 @olivecha
@ebezzam
Extending the sound could be added as an option in the future. It would be an interesting feature for time based analyses. Shortening was used because padding the array could add noise to the FFT and give the impression that it is more precise than it actually is.

I've made the suggested changes.
