### 2023-02-07 #4878 @olivecha
Thanks @ebezzam and @cwitkowitz, I've made the changes recommended by @ebezzam to the paper and fixed the failing test.
