+++ guitarsounds/__init__.py
@@ -3,4 +3,4 @@ from guitarsounds.interface import guitarGUI as Interface
 from guitarsounds.analysis import Signal, Sound, SoundPack, Plot
 
 # Version of the guitarsounds package
-__version__ = '1.1.0'
+__version__ = '1.2.0'
