### 2023-02-08 #4878 @olivecha
I've uploaded the JOSS article branch on Zenodo : DOI 10.5281/zenodo.7622901
And the main branch including all the changes mades to PyPi under version 1.2.0 : https://pypi.org/project/guitarsounds/
Thanks !
