### 2023-02-08 #4878 @danielskatz
@olivecha - I'm the track editor who will handle the rest of the processing.  Can you please change the title of the paper to remove the space before the ":"?  Can you also change the metadata in the Zenodo archive so that the title matches the title of the paper?  (If needed, see the section labeled "How can I edit the metadata of a published record?" in https://help.zenodo.org for how to do this.)

I'll also proofread the paper carefully, and may have more suggested changes.
