+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'guitarsounds : A Python package to visualize harmonic sounds for musical instrument design'
+title: 'guitarsounds: A Python package to visualize harmonic sounds for musical instrument design'
 tags:
   - Python
   - signal processing
@@ -29,11 +29,11 @@ Such sounds are defined as having a clear onset, and a frequency-amplitude distr
 Consequently, while the package is named `guitarsounds`, its analysis framework can be used with any harmonic sound, such as those produced by a piano, or a percussion instrument. 
 
 The `guitarsounds` package is divided in two main components. 
-First, the package is developed around a convenient object-oriented Application Programming Interface (API) which can be used to extract features from sounds and visualize them according to the user's needs. 
+First, the package is developed around a convenient object-oriented Application Programming Interface (API) that can be used to extract features from sounds and visualize them according to the user's needs. 
 Then, a graphical user interface (GUI) makes most of the features of `guitarsounds` available to users less knowledgeable in programming. 
-`guitarsounds` is meant to be used with the Jupyter Notebook interface as to allow interactively exploring the sound data, either with the API or the GUI.
+`guitarsounds` is meant to be used with the Jupyter Notebook interface to allow interactively exploring the sound data, either with the API or the GUI.
 
-The main features of `guitarsounds` are: 
+The main features of `guitarsounds` are to: 
 
 - Automate the loading, conditioning and normalization of multiple sound files to meaningfully compare their features.
 - Visualize sounds features relevant to musical instrument design, such as: 
@@ -45,12 +45,12 @@ The main features of `guitarsounds` are:
 - Extract numerical values for certain features such as the Helmholtz cavity frequency of a guitar.
 - Provide an easy-to-use signal processing API to compute new features meeting specific needs by providing access to lower level features and handling the differences between sound files, such as the file sample rate.
 
-Specifically, the API provides 4 classes nested together : 
+Specifically, the API provides four classes nested together: 
 
-- `Plot`: Low level plotting of specific features, such as plotting the Fast Fourier Transform (FFT) of a sound file.
-- `Signal`: Store the data of an array corresponding to a single signal. For example, if a sound file is read and filtered, the array resulting from the filtering operation will be stored in a new instance of the `Signal` class. An instance of the `Plot` class is constructed for each `Signal` class instance and stored as an attribute of the `Signal` class. The `Signal` class contains all the features relying only on the data of a single sound signal, such as the computation of the envelope.
-- `Sound`: Store all the information corresponding to a single sound file. When a `.wav` file is read using `guitarsounds`, all the processing is handled by the `Sound` class, such as truncating, filtering or normalizing the sound signal. The `Sound` class provides the features relying on more than one `Signal` instance, but still using the information from a single sound file, such as the power distribution of a sound across different frequency bands.
-- `SoundPack`: Contain multiple `Sound` class instances and provides the features used to compare the data between different sound files. The `SoundPack` methods are divided between methods developed to compare two sounds and methods developed to compare an arbitrary amount of sounds. As an example, the method plotting the FFT of two sounds in a mirror configuration can only be called if the `SoundPack` was constructed using exactly two sounds, whereas the method showing a table of the different sound fundamental frequencies can be called for a `SoundPack` instance created using an arbitrary number of `Sounds`.
+- `Plot`: Provides low level plotting of specific features, such as plotting the Fast Fourier Transform (FFT) of a sound file.
[…]
