### 2023-02-08 #4878 @danielskatz
Please see https://github.com/olivecha/guitarsounds/pull/48 for my suggested changes in the paper. Please merge this, or let me know what you disagree with, then we can continue the acceptance and publication
