### 2023-02-09 #4878 @olivecha
I've merged the suggested changes, I'll look into changing the metadata on zenodo.
