### 2023-02-09 #4878 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
