### 2023-02-09 #4878 @olivecha
I tried looking at the article on the JOSS website and the DOI does not seem to resolve. The editorialbot github action seems to have succeeded, I don't know if thats normal ?
